reorg config files

+++ script/local.sh
@@ -1,29 +1,39 @@
 #!/bin/bash
-
+if [ $# -lt 3 ]; then
+    echo "usage: ./local.sh num_servers num_workers app_conf [args]"
+    exit -1;
+fi
 dir=`dirname "$0"`
 cd ${dir}
 
 killall -q ps
 
-scheduler="role:SCHEDULER,hostname:'127.0.0.1',port:8000,id:'H'"
-W0="role:WORKER,hostname:'127.0.0.1',port:8001,id:'W0'"
-W1="role:WORKER,hostname:'127.0.0.1',port:8002,id:'W1'"
-W2="role:WORKER,hostname:'127.0.0.1',port:8003,id:'W2'"
-S0="role:SERVER,hostname:'127.0.0.1',port:8010,id:'S0'"
-S1="role:SERVER,hostname:'127.0.0.1',port:8011,id:'S1'"
-arg="-num_servers 1 -num_workers 1 -num_threads 1 -app ../config/online_rcv1_l1lr.conf"
+num_servers=$1
+shift
+num_workers=$1
+shift
+bin="../bin/ps"
+arg="-num_servers ${num_servers} -num_workers ${num_workers} -app $@"
 
 mkdir -p ../output
 FLAGS_logtostderr=1
 
-bin="../bin/ps"
+# start the scheduler
+Sch="role:SCHEDULER,hostname:'127.0.0.1',port:8000,id:'H'"
+${bin} -my_node ${Sch} -scheduler ${Sch} ${arg} &
 
-${bin} ${arg} -scheduler ${scheduler} -my_node ${scheduler} &
-${bin} ${arg} -scheduler ${scheduler} -my_node ${W0} &
-${bin} ${arg} -scheduler ${scheduler} -my_node ${S0} &
+# start servers
+for ((i=0; i<${num_servers}; ++i)); do
+    port=$((8100 + ${i}))
+    N="role:SERVER,hostname:'127.0.0.1',port:${port},id:'S${i}'"
+    ${bin} -my_node ${N} -scheduler ${Sch} ${arg} &
+done
 
-# ${bin} ${arg} -scheduler ${scheduler} -my_node ${W1} &
-# ${bin} ${arg} -scheduler ${scheduler} -my_node ${W2} &
-# ${bin} ${arg} -scheduler ${scheduler} -my_node ${S1} &
+# start workers
+for ((i=0; i<${num_workers}; ++i)); do
+    port=$((8200 + ${i}))
+    N="role:WORKER,hostname:'127.0.0.1',port:${port},id:'W${i}'"
+    ${bin} -my_node ${N} -scheduler ${Sch} ${arg} &
+done
 
 wait
